% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_measure.R
\name{cook_distance}
\alias{cook_distance}
\title{Cook's Distance Calculation}
\usage{
cook_distance(model)
}
\arguments{
\item{model}{An `lm` object representing the linear model.}
}
\value{
A vector of Cook's Distance values
}
\description{
Cook's Distance Calculation
}
