% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_measure.R
\name{dffits}
\alias{dffits}
\title{DFFITS Calculation}
\usage{
dffits(model)
}
\arguments{
\item{model}{An `lm` object representing the linear model.}
}
\value{
A vector of DFFITS values
}
\description{
DFFITS Calculation
}
