% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_measure.R
\name{hadi_influence}
\alias{hadi_influence}
\title{Hadi's Influence Measure Calculation}
\usage{
hadi_influence(model)
}
\arguments{
\item{model}{An `lm` object representing the linear model.}
}
\value{
A vector of Hadi's Influence Measure values
}
\description{
Hadi's Influence Measure Calculation
}
