% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_measure.R
\name{influence_measure_plot}
\alias{influence_measure_plot}
\title{Calculate Influence Measure and Plot the influence values}
\usage{
influence_measure_plot(data, model, method = c("cooks", "dffits", "hadi"))
}
\arguments{
\item{data}{A data frame containing the data.}

\item{model}{An `lm` object representing the linear model.}

\item{method}{A character string specifying the method of influence measure to calculate.
One of "cooks", "dffits", or "hadi".}
}
\value{
A scatter plot of the selected influence measure.
}
\description{
This function calculates one of Cook's Distance, DFFITS, or Hadi's Influence Measure
for a given linear model object and then creates a scatter plot of the influence values.
}
\examples{
lm_model <- lm(mpg ~ wt + hp, data = mtcars)
influence_measure_plot(mtcars, lm_model, method = "cooks")
influence_measure_plot(mtcars, lm_model, method = "dffits")
influence_measure_plot(mtcars, lm_model, method = "hadi")
}
